fix config SalesTaxControllerTest

+++ src/main/java/com/rsif/salestax/model/ReceiptItem.java
@@ -4,12 +4,10 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.math.BigDecimal;
-
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
 public class ReceiptItem {
     private String name;
-    private BigDecimal total;
+    private double total;
 }
